fix data type

+++ app/Database/Migrations/2024-02-24-165352_User.php
@@ -11,7 +11,7 @@ class User extends Migration
     {
         $fields = [
             'id' => [
-                'type' => 'INT',
+                'type' => 'BIGINT',
                 'constraint' => 5,
                 'unsigned' => true,
                 'auto_increment' => true,
